requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
billiard==4.1.0
black==23.7.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.3
django-wkhtmltopdf==3.4.0
djangorestframework==3.14.0
drf-spectacular==0.26.3
drf-yasg==1.21.6
flake8==6.0.0
idna==3.4
inflection==0.5.1
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
kombu==5.3.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
pdfkit==1.0.0
platformdirs==3.9.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==4.6.0
referencing==0.29.1
requests==2.31.0
rpds-py==0.8.10
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
